Move more painting related code out of lapce-data
