+++ lapce-data/src/editor.rs
@@ -9,22 +9,20 @@ use crate::command::CommandTarget;
 use crate::command::LapceCommandNew;
 use crate::command::LAPCE_NEW_COMMAND;
 use crate::completion::{CompletionData, CompletionStatus, Snippet};
-use crate::config::{Config, LapceTheme};
-use crate::data::EditorTabChild;
+use crate::config::Config;
 use crate::data::MotionMode;
 use crate::data::RegisterKind;
 use crate::data::{
     EditorDiagnostic, InlineFindDirection, LapceEditorData, LapceMainSplitData,
-    LapceTabData, RegisterData, SplitContent,
+    RegisterData, SplitContent,
 };
 use crate::movement::InsertDrift;
 use crate::proxy::path_from_url;
 use crate::state::LapceWorkspace;
-use crate::svg::get_svg;
 use crate::{buffer::WordProperty, movement::CursorMode};
 use crate::{
     command::{LapceCommand, LapceUICommand, LAPCE_UI_COMMAND},
-    movement::{ColPosition, Movement, SelRegion, Selection},
+    movement::{Movement, SelRegion, Selection},
     split::SplitMoveDirection,
     state::Mode,
     state::VisualMode,
@@ -34,15 +32,12 @@ use crate::{keypress::KeyPressFocus, movement::Cursor};
 use crate::{proxy::LapceProxy, source_control::SourceControlData};
 use anyhow::{anyhow, Result};
 use crossbeam_channel::{self, bounded};
-use druid::kurbo::BezPath;
+use druid::piet::PietTextLayout;
 use druid::piet::Svg;
-use druid::piet::{
-    PietTextLayout, Text, TextLayout as TextLayoutTrait, TextLayoutBuilder,
-};
 use druid::Modifiers;
 use druid::{
-    kurbo::Line, piet::PietText, Color, Command, Env, EventCtx, FontFamily,
-    PaintCtx, Point, Rect, RenderContext, Size, Target, Vec2, WidgetId,
+    piet::PietText, Command, Env, EventCtx, Point, Rect, Size, Target, Vec2,
+    WidgetId,
 };
 use druid::{Application, ExtEventSink, MouseEvent};
 pub use lapce_core::syntax::Syntax;
@@ -771,67 +766,6 @@ impl LapceEditorBufferData {
                 event_sink,
             );
         }
-    }
-
-    pub fn cursor_region(&self, text: &mut PietText, config: &Config) -> Rect {
-        let offset = self.editor.cursor.offset();
-        let (line, col) = self
-            .buffer
-            .offset_to_line_col(offset, self.config.editor.tab_width);
-        let width = config.editor_text_width(text, "W");
-        let cursor_x = col as f64 * width;
-        let line_height = config.editor.line_height as f64;
-
-        let y = if self.editor.code_lens {
-            let empty_vec = Vec::new();
-            let normal_lines = self
-                .buffer
-                .syntax
-                .as_ref()
-                .map(|s| &s.normal_lines)
-                .unwrap_or(&empty_vec);
-
-            let mut y = 0.0;
-            let mut current_line = 0;
-            let mut normal_lines = normal_lines.iter();
-            loop {
-                match normal_lines.next() {
-                    Some(next_normal_line) => {
-                        let next_normal_line = *next_normal_line;
-                        if next_normal_line < line {
-                            let chunk_height = config.editor.code_lens_font_size
-                                as f64
-                                * (next_normal_line - current_line) as f64
-                                + line_height;
-                            y += chunk_height;
-                            current_line = next_normal_line + 1;
-                            continue;
-                        };
-                        y += (line - current_line) as f64
-                            * config.editor.code_lens_font_size as f64;
-                        break;
-                    }
-                    None => {
-                        y += (line - current_line) as f64
-                            * config.editor.code_lens_font_size as f64;
-                        break;
-                    }
-                }
-            }
-            y
-        } else {
-            let line = if let Some(compare) = self.editor.compare.as_ref() {
-                self.buffer.diff_visual_line(compare, line)
-            } else {
-                line
-            };
-            line as f64 * line_height
-        };
-
-        Rect::ZERO
-            .with_size(Size::new(width, line_height))
-            .with_origin(Point::new(cursor_x, y))
-            .inflate(width, line_height)
     }
 
     pub fn update_global_search(&self, ctx: &mut EventCtx, pattern: String) {
@@ -1641,7 +1575,7 @@ impl LapceEditorBufferData {
         self.buffer.code_actions.get(&prev_offset)
     }
 
-    fn diagnostics(&self) -> Option<&Arc<Vec<EditorDiagnostic>>> {
+    pub fn diagnostics(&self) -> Option<&Arc<Vec<EditorDiagnostic>>> {
         if let BufferContent::File(path) = &self.buffer.content {
             self.main_split.diagnostics.get(path)
         } else {
@@ -1649,646 +1583,11 @@ impl LapceEditorBufferData {
         }
     }
 
-    fn diagnostics_mut(&mut self) -> Option<&mut Vec<EditorDiagnostic>> {
+    pub fn diagnostics_mut(&mut self) -> Option<&mut Vec<EditorDiagnostic>> {
         if let BufferContent::File(path) = &self.buffer.content {
             self.main_split.diagnostics.get_mut(path).map(Arc::make_mut)
         } else {
             None
-        }
-    }
-
-    pub fn paint_code_lens_content(
-        &self,
-        ctx: &mut PaintCtx,
-        is_focused: bool,
-        config: &Config,
-    ) {
-        let rect = ctx.region().bounding_box();
-
-        let ref_text_layout = ctx
-            .text()
-            .new_text_layout("W")
-            .font(
-                self.config.editor.font_family(),
-                self.config.editor.font_size as f64,
-            )
-            .build()
-            .unwrap();
-        let char_width = ref_text_layout.size().width;
-        let y_shift =
-            (config.editor.line_height as f64 - ref_text_layout.size().height) / 2.0;
-        let small_char_width =
-            config.char_width(ctx.text(), config.editor.code_lens_font_size as f64);
-
-        let empty_lens = Syntax::lens_from_normal_lines(
-            self.buffer.len(),
-            config.editor.line_height,
-            config.editor.code_lens_font_size,
-            &[],
-        );
-        let lens = if let Some(syntax) = self.buffer.syntax.as_ref() {
-            &syntax.lens
-        } else {
-            &empty_lens
-        };
-
-        let cursor_line = self
-            .buffer
-            .line_of_offset(self.editor.cursor.offset().min(self.buffer.len()));
-        let last_line = self.buffer.line_of_offset(self.buffer.len());
-        let start_line =
-            lens.line_of_height(rect.y0.floor() as usize).min(last_line);
-        let end_line = lens
-            .line_of_height(rect.y1.ceil() as usize + config.editor.line_height)
-            .min(last_line);
-        let start_offset = self.buffer.offset_of_line(start_line);
-        let end_offset = self.buffer.offset_of_line(end_line + 1);
-        let mut lines_iter = self.buffer.rope.lines(start_offset..end_offset);
-
-        let mut y = lens.height_of_line(start_line) as f64;
-        for (line, line_height) in lens.iter_chunks(start_line..end_line + 1) {
-            if let Some(line_content) = lines_iter.next() {
-                let is_small = line_height < config.editor.line_height;
-
-                let mut x = 0.0;
-                if is_small {
-                    for ch in line_content.chars() {
-                        if ch == ' ' {
-                            x += char_width - small_char_width;
-                        } else if ch == '\t' {
-                            x += (char_width - small_char_width)
-                                * config.editor.tab_width as f64;
-                        } else {
-                            break;
-                        }
-                    }
-                }
-
-                self.paint_cursor_on_line(
-                    ctx,
-                    is_focused,
-                    cursor_line,
-                    line,
-                    x,
-                    y,
-                    if is_small {
-                        small_char_width
-                    } else {
-                        char_width
-                    },
-                    line_height as f64,
-                    config,
-                );
-                let text_layout = self.buffer.new_text_layout(
-                    ctx,
-                    line,
-                    &line_content,
-                    None,
-                    if is_small {
-                        config.editor.code_lens_font_size
-                    } else {
-                        config.editor.font_size
-                    },
-                    [rect.x0, rect.x1],
-                    config,
-                );
-                ctx.draw_text(
-                    &text_layout,
-                    Point::new(x, if is_small { y } else { y + y_shift }),
-                );
-                y += line_height as f64;
-            }
-        }
-    }
-
-    pub fn paint_content(
-        &self,
-        ctx: &mut PaintCtx,
-        is_focused: bool,
-        placeholder: Option<&String>,
-        config: &Config,
-        env: &Env,
-    ) {
-        let line_height = self.line_height(env);
-        let line_padding = self.line_padding(env);
-
-        let font_size = if self.editor.content.is_input() {
-            env.get(LapceTheme::INPUT_FONT_SIZE) as usize
-        } else {
-            self.config.editor.font_size
-        };
-
-        let text_layout = ctx
-            .text()
-            .new_text_layout("W")
-            .font(self.config.editor.font_family(), font_size as f64)
-            .build()
-            .unwrap();
-        let char_width = text_layout.size().width;
-        let y_shift = (line_height - text_layout.size().height) / 2.0;
-
-        if self.editor.content.is_input()
-            || (self.editor.compare.is_none() && !self.editor.code_lens)
-        {
-            self.paint_cursor(ctx, is_focused, placeholder, char_width, config, env);
-            self.paint_find(ctx, char_width, env);
-        }
-        let self_size = ctx.size();
-        let rect = ctx.region().bounding_box();
-        let start_line = (rect.y0 / line_height).floor() as usize;
-        let end_line = (rect.y1 / line_height).ceil() as usize;
-
-        if !self.editor.content.is_input() && self.editor.code_lens {
-            self.paint_code_lens_content(ctx, is_focused, config);
-        } else if let Some(compare) = self.editor.compare.as_ref() {
-            if let Some(changes) = self.buffer.history_changes.get(compare) {
-                let cursor_line =
-                    self.buffer.line_of_offset(self.editor.cursor.offset());
-                let mut line = 0;
-                for change in changes.iter() {
-                    match change {
-                        DiffLines::Left(range) => {
-                            let len = range.len();
-                            line += len;
-
-                            if line < start_line {
-                                continue;
-                            }
-                            ctx.fill(
-                                Size::new(self_size.width, line_height * len as f64)
-                                    .to_rect()
-                                    .with_origin(Point::new(
-                                        0.0,
-                                        line_height * (line - len) as f64,
-                                    )),
-                                config.get_color_unchecked(
-                                    LapceTheme::SOURCE_CONTROL_REMOVED,
-                                ),
-                            );
-                            for l in line - len..line {
-                                if l < start_line {
-                                    continue;
-                                }
-                                let actual_line = l - (line - len) + range.start;
-                                if let Some(text_layout) =
-                                    self.buffer.history_text_layout(
-                                        ctx,
-                                        compare,
-                                        actual_line,
-                                        None,
-                                        [rect.x0, rect.x1],
-                                        config,
-                                    )
-                                {
-                                    ctx.draw_text(
-                                        &text_layout,
-                                        Point::new(
-                                            0.0,
-                                            line_height * l as f64 + y_shift,
-                                        ),
-                                    );
-                                }
-                                if l > end_line {
-                                    break;
-                                }
-                            }
-                        }
-                        DiffLines::Skip(left, right) => {
-                            let rect = Size::new(self_size.width, line_height)
-                                .to_rect()
-                                .with_origin(Point::new(
-                                    0.0,
-                                    line_height * line as f64,
-                                ));
-                            ctx.fill(
-                                rect,
-                                self.config.get_color_unchecked(
-                                    LapceTheme::PANEL_BACKGROUND,
-                                ),
-                            );
-                            ctx.stroke(
-                                rect,
-                                config.get_color_unchecked(
-                                    LapceTheme::EDITOR_FOREGROUND,
-                                ),
-                                1.0,
-                            );
-                            let text_layout = ctx
-                                .text()
-                                .new_text_layout(format!(
-                                    " -{}, +{}",
-                                    left.end + 1,
-                                    right.end + 1
-                                ))
-                                .font(config.editor.font_family(), font_size as f64)
-                                .text_color(
-                                    config
-                                        .get_color_unchecked(
-                                            LapceTheme::EDITOR_FOREGROUND,
-                                        )
-                                        .clone(),
-                                )
-                                .build_with_info(
-                                    true,
-                                    config.editor.tab_width,
-                                    Some([rect.x0, rect.x1]),
-                                );
-                            ctx.draw_text(
-                                &text_layout,
-                                Point::new(0.0, line_height * line as f64 + y_shift),
-                            );
-                            line += 1;
-                        }
-                        DiffLines::Both(_left, right) => {
-                            let len = right.len();
-                            line += len;
-                            if line < start_line {
-                                continue;
-                            }
-                            for l in line - len..line {
-                                if l < start_line {
-                                    continue;
-                                }
-                                let rope_line = l - (line - len) + right.start;
-                                self.paint_cursor_on_line(
-                                    ctx,
-                                    is_focused,
-                                    cursor_line,
-                                    rope_line,
-                                    0.0,
-                                    l as f64 * line_height,
-                                    char_width,
-                                    line_height,
-                                    config,
-                                );
-                                let text_layout = self.buffer.new_text_layout(
-                                    ctx,
-                                    rope_line,
-                                    &self.buffer.line_content(rope_line),
-                                    None,
-                                    font_size,
-                                    [rect.x0, rect.x1],
-                                    &self.config,
-                                );
-                                ctx.draw_text(
-                                    &text_layout,
-                                    Point::new(
-                                        0.0,
-                                        line_height * l as f64 + y_shift,
-                                    ),
-                                );
-                                if l > end_line {
-                                    break;
-                                }
-                            }
-                        }
-                        DiffLines::Right(range) => {
-                            let len = range.len();
-                            line += len;
-
-                            if line < start_line {
-                                continue;
-                            }
-
-                            ctx.fill(
-                                Size::new(
-                                    self_size.width,
-                                    line_height * range.len() as f64,
-                                )
-                                .to_rect()
-                                .with_origin(
-                                    Point::new(
-                                        0.0,
-                                        line_height * (line - range.len()) as f64,
-                                    ),
-                                ),
-                                config.get_color_unchecked(
-                                    LapceTheme::SOURCE_CONTROL_ADDED,
-                                ),
-                            );
-
-                            for l in line - len..line {
-                                if l < start_line {
-                                    continue;
-                                }
-                                let rope_line = l - (line - len) + range.start;
-                                self.paint_cursor_on_line(
-                                    ctx,
-                                    is_focused,
-                                    cursor_line,
-                                    rope_line,
-                                    0.0,
-                                    l as f64 * line_height,
-                                    char_width,
-                                    line_height,
-                                    config,
-                                );
-                                let text_layout = self.buffer.new_text_layout(
-                                    ctx,
-                                    rope_line,
-                                    &self.buffer.line_content(rope_line),
-                                    None,
-                                    font_size,
-                                    [rect.x0, rect.x1],
-                                    &self.config,
-                                );
-                                ctx.draw_text(
-                                    &text_layout,
-                                    Point::new(
-                                        0.0,
-                                        line_height * l as f64 + y_shift,
-                                    ),
-                                );
-                                if l > end_line {
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            return;
-        } else {
-            let cursor_offset = self.editor.cursor.offset();
-            let cursor_line = self.buffer.line_of_offset(cursor_offset);
-            let start_offset = self.buffer.offset_of_line(start_line);
-            let end_offset = self.buffer.offset_of_line(end_line + 1);
-            let mode = self.editor.cursor.get_mode();
-            for (i, line_content) in self
-                .buffer
-                .slice_to_cow(start_offset..end_offset)
-                .split('\n')
-                .enumerate()
-            {
-                let line = i + start_line;
-                let cursor_index =
-                    if is_focused && mode != Mode::Insert && line == cursor_line {
-                        let cursor_line_start = self
-                            .buffer
-                            .offset_of_line(cursor_line)
-                            .min(self.buffer.len());
-                        let index = self
-                            .buffer
-                            .slice_to_cow(cursor_line_start..cursor_offset)
-                            .len();
-                        Some(index)
-                    } else {
-                        None
-                    };
-                let text_layout = self.buffer.new_text_layout(
-                    ctx,
-                    line,
-                    line_content,
-                    cursor_index,
-                    font_size,
-                    [rect.x0, rect.x1],
-                    &self.config,
-                );
-                ctx.draw_text(
-                    &text_layout,
-                    Point::new(
-                        0.0,
-                        line_height * line as f64 + y_shift + line_padding,
-                    ),
-                );
-            }
-        }
-
-        self.paint_snippet(ctx);
-        self.paint_diagnostics(ctx);
-        if self.buffer.len() == 0 {
-            if let Some(placeholder) = placeholder {
-                let text_layout = ctx
-                    .text()
-                    .new_text_layout(placeholder.to_string())
-                    .font(FontFamily::SYSTEM_UI, 13.0)
-                    .text_color(
-                        self.config
-                            .get_color_unchecked(LapceTheme::EDITOR_DIM)
-                            .clone(),
-                    )
-                    .build()
-                    .unwrap();
-                ctx.draw_text(&text_layout, Point::new(0.0, y_shift));
-            }
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn paint_cursor_on_line(
-        &self,
-        ctx: &mut PaintCtx,
-        is_focused: bool,
-        cursor_line: usize,
-        actual_line: usize,
-        x_shift: f64,
-        y: f64,
-        char_width: f64,
-        line_height: f64,
-        config: &Config,
-    ) {
-        match &self.editor.cursor.mode {
-            CursorMode::Normal(_) => {}
-            CursorMode::Visual { start, end, mode } => {
-                let (start_line, start_col) = self.buffer.offset_to_line_col(
-                    *start.min(end),
-                    self.config.editor.tab_width,
-                );
-                let (end_line, end_col) = self.buffer.offset_to_line_col(
-                    *start.max(end),
-                    self.config.editor.tab_width,
-                );
-                if actual_line < start_line || actual_line > end_line {
-                    return;
-                }
-
-                let left_col = match mode {
-                    VisualMode::Normal => {
-                        if start_line == actual_line {
-                            start_col
-                        } else {
-                            0
-                        }
-                    }
-                    VisualMode::Linewise => 0,
-                    VisualMode::Blockwise => {
-                        let max_col = self.buffer.line_end_col(
-                            actual_line,
-                            false,
-                            self.config.editor.tab_width,
-                        );
-                        let left = start_col.min(end_col);
-                        if left > max_col {
-                            return;
-                        }
-                        left
-                    }
-                };
-
-                let right_col = match mode {
-                    VisualMode::Normal => {
-                        if actual_line == end_line {
-                            let max_col = self.buffer.line_end_col(
-                                actual_line,
-                                true,
-                                self.config.editor.tab_width,
-                            );
-                            (end_col + 1).min(max_col)
-                        } else {
-                            self.buffer.line_end_col(
-                                actual_line,
-                                true,
-                                self.config.editor.tab_width,
-                            ) + 1
-                        }
-                    }
-                    VisualMode::Linewise => {
-                        self.buffer.line_end_col(
-                            actual_line,
-                            true,
-                            self.config.editor.tab_width,
-                        ) + 1
-                    }
-                    VisualMode::Blockwise => {
-                        let max_col = self.buffer.line_end_col(
-                            actual_line,
-                            true,
-                            self.config.editor.tab_width,
-                        );
-                        let right = match self.editor.cursor.horiz.as_ref() {
-                            Some(&ColPosition::End) => max_col,
-                            _ => (end_col.max(start_col) + 1).min(max_col),
-                        };
-                        right
-                    }
-                };
-
-                let x0 = left_col as f64 * char_width + x_shift;
-                let x1 = right_col as f64 * char_width + x_shift;
-                let y0 = y;
-                let y1 = y0 + line_height;
-                ctx.fill(
-                    Rect::new(x0, y0, x1, y1),
-                    self.config
-                        .get_color_unchecked(LapceTheme::EDITOR_SELECTION),
-                );
-            }
-            CursorMode::Insert(selection) => {
-                let start_offset = self.buffer.offset_of_line(actual_line);
-                let end_offset = self.buffer.offset_of_line(actual_line + 1);
-                let regions = selection.regions_in_range(start_offset, end_offset);
-                for region in regions {
-                    if region.is_caret() {
-                        let caret_actual_line =
-                            self.buffer.line_of_offset(region.end());
-                        if caret_actual_line == actual_line {
-                            let size = ctx.size();
-                            ctx.fill(
-                                Rect::ZERO
-                                    .with_origin(Point::new(0.0, y))
-                                    .with_size(Size::new(size.width, line_height)),
-                                self.config.get_color_unchecked(
-                                    LapceTheme::EDITOR_CURRENT_LINE,
-                                ),
-                            );
-                        }
-                    } else {
-                        let start = region.start();
-                        let end = region.end();
-                        let (start_line, start_col) =
-                            self.buffer.offset_to_line_col(
-                                start.min(end),
-                                self.config.editor.tab_width,
-                            );
-                        let (end_line, end_col) = self.buffer.offset_to_line_col(
-                            start.max(end),
-                            self.config.editor.tab_width,
-                        );
-                        let left_col = match actual_line {
-                            _ if actual_line == start_line => start_col,
-                            _ => 0,
-                        };
-                        let right_col = match actual_line {
-                            _ if actual_line == end_line => {
-                                let max_col = self.buffer.line_end_col(
-                                    actual_line,
-                                    true,
-                                    self.config.editor.tab_width,
-                                );
-                                end_col.min(max_col)
-                            }
-                            _ => self.buffer.line_end_col(
-                                actual_line,
-                                true,
-                                self.config.editor.tab_width,
-                            ),
-                        };
-                        let x0 = left_col as f64 * char_width + x_shift;
-                        let x1 = right_col as f64 * char_width + x_shift;
-                        let y0 = y;
-                        let y1 = y0 + line_height;
-                        ctx.fill(
-                            Rect::new(x0, y0, x1, y1),
-                            self.config
-                                .get_color_unchecked(LapceTheme::EDITOR_SELECTION),
-                        );
-                    }
-                }
-                for region in regions {
-                    if is_focused {
-                        let (caret_actual_line, col) =
-                            self.buffer.offset_to_line_col(
-                                region.end(),
-                                self.config.editor.tab_width,
-                            );
-                        if caret_actual_line == actual_line {
-                            let x = col as f64 * char_width + x_shift;
-                            ctx.stroke(
-                                Line::new(
-                                    Point::new(x, y),
-                                    Point::new(x, y + line_height),
-                                ),
-                                self.config
-                                    .get_color_unchecked(LapceTheme::EDITOR_CARET),
-                                2.0,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-        if cursor_line == actual_line {
-            if let CursorMode::Normal(_) = &self.editor.cursor.mode {
-                let size = ctx.size();
-                ctx.fill(
-                    Rect::ZERO
-                        .with_origin(Point::new(0.0, y))
-                        .with_size(Size::new(size.width, line_height)),
-                    self.config
-                        .get_color_unchecked(LapceTheme::EDITOR_CURRENT_LINE),
-                );
-            }
-            match &self.editor.cursor.mode {
-                CursorMode::Normal(_) | CursorMode::Visual { .. } => {
-                    if is_focused {
-                        let (x0, x1) = self.editor.cursor.current_char(
-                            &self.buffer,
-                            char_width,
-                            config,
-                        );
-                        let cursor_width =
-                            if x1 > x0 { x1 - x0 } else { char_width };
-                        ctx.fill(
-                            Rect::ZERO
-                                .with_origin(Point::new(x0 + x_shift, y))
-                                .with_size(Size::new(cursor_width, line_height)),
-                            self.config
-                                .get_color_unchecked(LapceTheme::EDITOR_CARET),
-                        );
-                    }
-                }
-                CursorMode::Insert(_) => {}
-            }
         }
     }
 
@@ -2432,608 +1731,6 @@ impl LapceEditorBufferData {
             mouse_event.mods.shift(),
             mouse_event.mods.alt(),
         );
-    }
-
-    fn line_height(&self, env: &Env) -> f64 {
-        if self.editor.content.is_input() {
-            env.get(LapceTheme::INPUT_LINE_HEIGHT)
-        } else {
-            self.config.editor.line_height as f64
-        }
-    }
-
-    fn line_padding(&self, env: &Env) -> f64 {
-        if self.editor.content.is_input() {
-            env.get(LapceTheme::INPUT_LINE_PADDING)
-        } else {
-            0.0
-        }
-    }
-
-    fn paint_cursor(
-        &self,
-        ctx: &mut PaintCtx,
-        is_focused: bool,
-        placeholder: Option<&String>,
-        width: f64,
-        config: &Config,
-        env: &Env,
-    ) {
-        let line_height = self.line_height(env);
-        let line_padding = self.line_padding(env);
-        let start_line =
-            (self.editor.scroll_offset.y / line_height).floor() as usize;
-        let end_line = ((self.editor.size.borrow().height
-            + self.editor.scroll_offset.y)
-            / line_height)
-            .ceil() as usize;
-        match &self.editor.cursor.mode {
-            CursorMode::Normal(offset) => {
-                let line = self.buffer.line_of_offset(*offset);
-                self.paint_cursor_line(ctx, line, is_focused, placeholder);
-
-                if is_focused {
-                    let (x0, x1) =
-                        self.editor.cursor.current_char(&self.buffer, width, config);
-                    let char_width = if x1 > x0 { x1 - x0 } else { width };
-                    ctx.fill(
-                        Rect::ZERO
-                            .with_origin(Point::new(
-                                x0,
-                                line as f64 * line_height + line_padding,
-                            ))
-                            .with_size(Size::new(char_width, line_height)),
-                        self.config.get_color_unchecked(LapceTheme::EDITOR_CARET),
-                    );
-                }
-            }
-            CursorMode::Visual { start, end, mode } => {
-                let paint_start_line = start_line;
-                let paint_end_line = end_line;
-                let (start_line, start_col) = self.buffer.offset_to_line_col(
-                    *start.min(end),
-                    self.config.editor.tab_width,
-                );
-                let (end_line, end_col) = self.buffer.offset_to_line_col(
-                    *start.max(end),
-                    self.config.editor.tab_width,
-                );
-                for line in paint_start_line..paint_end_line {
-                    if line < start_line || line > end_line {
-                        continue;
-                    }
-                    let line_content = self.buffer.line_content(line);
-                    let left_col = match mode {
-                        VisualMode::Normal => match line {
-                            _ if line == start_line => start_col,
-                            _ => 0,
-                        },
-                        VisualMode::Linewise => 0,
-                        VisualMode::Blockwise => {
-                            let max_col = self.buffer.line_end_col(
-                                line,
-                                false,
-                                self.config.editor.tab_width,
-                            );
-                            let left = start_col.min(end_col);
-                            if left > max_col {
-                                continue;
-                            }
-                            left
-                        }
-                    };
-                    let x0 = left_col as f64 * width;
-
-                    let right_col = match mode {
-                        VisualMode::Normal => match line {
-                            _ if line == end_line => {
-                                let max_col = self.buffer.line_end_col(
-                                    line,
-                                    true,
-                                    self.config.editor.tab_width,
-                                );
-                                (end_col + 1).min(max_col)
-                            }
-                            _ => {
-                                self.buffer.line_end_col(
-                                    line,
-                                    true,
-                                    self.config.editor.tab_width,
-                                ) + 1
-                            }
-                        },
-                        VisualMode::Linewise => {
-                            self.buffer.line_end_col(
-                                line,
-                                true,
-                                self.config.editor.tab_width,
-                            ) + 1
-                        }
-                        VisualMode::Blockwise => {
-                            let max_col = self.buffer.line_end_col(
-                                line,
-                                true,
-                                self.config.editor.tab_width,
-                            );
-                            let right = match self.editor.cursor.horiz.as_ref() {
-                                Some(&ColPosition::End) => max_col,
-                                _ => (end_col.max(start_col) + 1).min(max_col),
-                            };
-                            right
-                        }
-                    };
-                    if !line_content.is_empty() {
-                        let x1 = right_col as f64 * width;
-
-                        let y0 = line as f64 * line_height + line_padding;
-                        let y1 = y0 + line_height;
-                        ctx.fill(
-                            Rect::new(x0, y0, x1, y1),
-                            self.config
-                                .get_color_unchecked(LapceTheme::EDITOR_SELECTION),
-                        );
-                    }
-
-                    if is_focused {
-                        let line = self.buffer.line_of_offset(*end);
-
-                        let (x0, x1) = self.editor.cursor.current_char(
-                            &self.buffer,
-                            width,
-                            config,
-                        );
-                        let char_width = if x1 > x0 { x1 - x0 } else { width };
-                        ctx.fill(
-                            Rect::ZERO
-                                .with_origin(Point::new(
-                                    x0,
-                                    line as f64 * line_height + line_padding,
-                                ))
-                                .with_size(Size::new(char_width, line_height)),
-                            self.config
-                                .get_color_unchecked(LapceTheme::EDITOR_CARET),
-                        );
-                    }
-                }
-            }
-            CursorMode::Insert(selection) => {
-                let offset = selection.get_cursor_offset();
-                let _line = self.buffer.line_of_offset(offset);
-                let last_line = self.buffer.last_line();
-                let end_line = if end_line > last_line {
-                    last_line
-                } else {
-                    end_line
-                };
-                let start = self.buffer.offset_of_line(start_line);
-                let end = self.buffer.offset_of_line(end_line + 1);
-                let regions = selection.regions_in_range(start, end);
-                for region in regions {
-                    if region.start() == region.end() {
-                        let line = self.buffer.line_of_offset(region.start());
-                        self.paint_cursor_line(ctx, line, is_focused, placeholder);
-                    } else {
-                        let start = region.start();
-                        let end = region.end();
-                        let paint_start_line = start_line;
-                        let paint_end_line = end_line;
-                        let (start_line, start_col) =
-                            self.buffer.offset_to_line_col(
-                                start.min(end),
-                                self.config.editor.tab_width,
-                            );
-                        let (end_line, end_col) = self.buffer.offset_to_line_col(
-                            start.max(end),
-                            self.config.editor.tab_width,
-                        );
-                        for line in paint_start_line..paint_end_line + 1 {
-                            if line < start_line || line > end_line {
-                                continue;
-                            }
-
-                            let line_content = self.buffer.line_content(line);
-                            let left_col = match line {
-                                _ if line == start_line => start_col,
-                                _ => 0,
-                            };
-                            let x0 = left_col as f64 * width;
-
-                            let right_col = match line {
-                                _ if line == end_line => {
-                                    let max_col = self.buffer.line_end_col(
-                                        line,
-                                        true,
-                                        self.config.editor.tab_width,
-                                    );
-                                    end_col.min(max_col)
-                                }
-                                _ => self.buffer.line_end_col(
-                                    line,
-                                    true,
-                                    self.config.editor.tab_width,
-                                ),
-                            };
-
-                            if !line_content.is_empty() {
-                                let x1 = right_col as f64 * width;
-                                let y0 = line as f64 * line_height + line_padding;
-                                let y1 = y0 + line_height;
-                                ctx.fill(
-                                    Rect::new(x0, y0, x1, y1),
-                                    self.config.get_color_unchecked(
-                                        LapceTheme::EDITOR_SELECTION,
-                                    ),
-                                );
-                            }
-                        }
-                    }
-                }
-
-                for region in regions {
-                    if is_focused {
-                        let (line, col) = self.buffer.offset_to_line_col(
-                            region.end(),
-                            self.config.editor.tab_width,
-                        );
-                        let x = col as f64 * width;
-                        let y = line as f64 * line_height + line_padding;
-                        ctx.stroke(
-                            Line::new(
-                                Point::new(x, y),
-                                Point::new(x, y + line_height),
-                            ),
-                            self.config
-                                .get_color_unchecked(LapceTheme::EDITOR_CARET),
-                            2.0,
-                        )
-                    }
-                }
-            }
-        }
-    }
-
-    fn paint_cursor_line(
-        &self,
-        ctx: &mut PaintCtx,
-        line: usize,
-        is_focused: bool,
-        placeholder: Option<&String>,
-    ) {
-        if !is_focused && self.buffer.len() == 0 && placeholder.is_some() {
-            return;
-        }
-        if self.editor.content.is_input() {
-            return;
-        }
-        let line_height = self.config.editor.line_height as f64;
-        let size = ctx.size();
-        ctx.fill(
-            Rect::ZERO
-                .with_origin(Point::new(0.0, line as f64 * line_height))
-                .with_size(Size::new(size.width, line_height)),
-            self.config
-                .get_color_unchecked(LapceTheme::EDITOR_CURRENT_LINE),
-        );
-    }
-
-    fn paint_find(&self, ctx: &mut PaintCtx, char_width: f64, env: &Env) {
-        if self.editor.content.is_search() {
-            return;
-        }
-        if !self.find.visual {
-            return;
-        }
-        let line_height = self.line_height(env);
-        let start_line =
-            (self.editor.scroll_offset.y / line_height).floor() as usize;
-        let end_line = ((self.editor.size.borrow().height
-            + self.editor.scroll_offset.y)
-            / line_height)
-            .ceil() as usize;
-        let start_offset = self.buffer.offset_of_line(start_line);
-        let end_offset = self.buffer.offset_of_line(end_line + 1);
-        let cursor_offset = self.editor.cursor.offset();
-
-        self.buffer.update_find(&self.find, start_line, end_line);
-        if self.find.search_string.is_some() {
-            for region in self
-                .buffer
-                .find
-                .borrow()
-                .occurrences()
-                .regions_in_range(start_offset, end_offset)
-            {
-                let start = region.min();
-                let end = region.max();
-                let active = start <= cursor_offset && cursor_offset <= end;
-                let (start_line, start_col) = self
-                    .buffer
-                    .offset_to_line_col(start, self.config.editor.tab_width);
-                let (end_line, end_col) = self
-                    .buffer
-                    .offset_to_line_col(end, self.config.editor.tab_width);
-                for line in start_line..end_line + 1 {
-                    let left_col = if line == start_line { start_col } else { 0 };
-                    let right_col = if line == end_line {
-                        end_col
-                    } else {
-                        self.buffer.line_end_col(
-                            line,
-                            true,
-                            self.config.editor.tab_width,
-                        ) + 1
-                    };
-                    let x0 = left_col as f64 * char_width;
-                    let x1 = right_col as f64 * char_width;
-                    let y0 = line as f64 * line_height;
-                    let y1 = y0 + line_height;
-                    let rect = Rect::new(x0, y0, x1, y1);
-                    if active {
-                        ctx.fill(
-                            rect,
-                            &self
-                                .config
-                                .get_color_unchecked(LapceTheme::EDITOR_CARET)
-                                .clone()
-                                .with_alpha(0.5),
-                        );
-                    }
-                    ctx.stroke(
-                        rect,
-                        self.config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND),
-                        1.0,
-                    );
-                }
-            }
-        }
-    }
-
-    fn paint_snippet(&self, ctx: &mut PaintCtx) {
-        let line_height = self.config.editor.line_height as f64;
-        let start_line =
-            (self.editor.scroll_offset.y / line_height).floor() as usize;
-        let end_line = ((self.editor.size.borrow().height
-            + self.editor.scroll_offset.y)
-            / line_height)
-            .ceil() as usize;
-        let width = self.config.editor_text_width(ctx.text(), "W");
-        if let Some(snippet) = self.editor.snippet.as_ref() {
-            for (_, (start, end)) in snippet {
-                let paint_start_line = start_line;
-                let paint_end_line = end_line;
-                let (start_line, start_col) = self.buffer.offset_to_line_col(
-                    *start.min(end),
-                    self.config.editor.tab_width,
-                );
-                let (end_line, end_col) = self.buffer.offset_to_line_col(
-                    *start.max(end),
-                    self.config.editor.tab_width,
-                );
-                for line in paint_start_line..paint_end_line {
-                    if line < start_line || line > end_line {
-                        continue;
-                    }
-                    let line_content = self.buffer.line_content(line);
-                    let left_col = match line {
-                        _ if line == start_line => start_col,
-                        _ => 0,
-                    };
-                    let x0 = left_col as f64 * width;
-
-                    let right_col = match line {
-                        _ if line == end_line => {
-                            let max_col = self.buffer.line_end_col(
-                                line,
-                                true,
-                                self.config.editor.tab_width,
-                            );
-                            end_col.min(max_col)
-                        }
-                        _ => self.buffer.line_end_col(
-                            line,
-                            true,
-                            self.config.editor.tab_width,
-                        ),
-                    };
-                    if !line_content.is_empty() {
-                        let x1 = right_col as f64 * width;
-                        let y0 = line as f64 * line_height;
-                        let y1 = y0 + line_height;
-                        ctx.stroke(
-                            Rect::new(x0, y0, x1, y1).inflate(1.0, -0.5),
-                            self.config
-                                .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND),
-                            1.0,
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    fn paint_diagnostics(&self, ctx: &mut PaintCtx) {
-        let line_height = self.config.editor.line_height as f64;
-        let start_line =
-            (self.editor.scroll_offset.y / line_height).floor() as usize;
-        let end_line = ((self.editor.size.borrow().height
-            + self.editor.scroll_offset.y)
-            / line_height)
-            .ceil() as usize;
-
-        let width = self.config.editor_text_width(ctx.text(), "W");
-        let mut current = None;
-        let cursor_offset = self.editor.cursor.offset();
-        if let Some(diagnostics) = self.diagnostics() {
-            for diagnostic in diagnostics.iter() {
-                let start = diagnostic.diagnositc.range.start;
-                let end = diagnostic.diagnositc.range.end;
-                if (start.line as usize) <= end_line
-                    && (end.line as usize) >= start_line
-                {
-                    let start_offset = if let Some(range) = diagnostic.range {
-                        range.0
-                    } else {
-                        self.buffer
-                            .offset_of_position(&start, self.config.editor.tab_width)
-                    };
-                    if start_offset == cursor_offset {
-                        current = Some(diagnostic.clone());
-                    }
-                    for line in start.line as usize..end.line as usize + 1 {
-                        if line < start_line {
-                            continue;
-                        }
-                        if line > end_line {
-                            break;
-                        }
-
-                        let x0 = if line == start.line as usize {
-                            start.character as f64 * width
-                        } else {
-                            let (_, col) = self.buffer.offset_to_line_col(
-                                self.buffer.first_non_blank_character_on_line(line),
-                                self.config.editor.tab_width,
-                            );
-                            col as f64 * width
-                        };
-                        let x1 = if line == end.line as usize {
-                            end.character as f64 * width
-                        } else {
-                            (self.buffer.line_end_col(
-                                line,
-                                false,
-                                self.config.editor.tab_width,
-                            ) + 1) as f64
-                                * width
-                        };
-                        let _y1 = (line + 1) as f64 * line_height;
-                        let y0 = (line + 1) as f64 * line_height - 4.0;
-
-                        let severity = diagnostic
-                            .diagnositc
-                            .severity
-                            .as_ref()
-                            .unwrap_or(&DiagnosticSeverity::Information);
-                        let color = match severity {
-                            DiagnosticSeverity::Error => self
-                                .config
-                                .get_color_unchecked(LapceTheme::LAPCE_ERROR),
-                            DiagnosticSeverity::Warning => self
-                                .config
-                                .get_color_unchecked(LapceTheme::LAPCE_WARN),
-                            _ => self
-                                .config
-                                .get_color_unchecked(LapceTheme::LAPCE_WARN),
-                        };
-                        paint_wave_line(ctx, Point::new(x0, y0), x1 - x0, color);
-                    }
-                }
-            }
-        }
-
-        if let Some(diagnostic) = current {
-            if self.editor.cursor.is_normal() {
-                let text_layout = ctx
-                    .text()
-                    .new_text_layout(diagnostic.diagnositc.message.clone())
-                    .font(FontFamily::SYSTEM_UI, 14.0)
-                    .text_color(
-                        self.config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND)
-                            .clone(),
-                    )
-                    .max_width(self.editor.size.borrow().width - 20.0)
-                    .build()
-                    .unwrap();
-                let text_size = text_layout.size();
-                let mut text_height = text_size.height;
-
-                let related = diagnostic
-                    .diagnositc
-                    .related_information
-                    .map(|related| {
-                        related
-                            .iter()
-                            .map(|i| {
-                                let text_layout = ctx
-                                    .text()
-                                    .new_text_layout(i.message.clone())
-                                    .font(FontFamily::SYSTEM_UI, 14.0)
-                                    .text_color(
-                                        self.config
-                                            .get_color_unchecked(
-                                                LapceTheme::EDITOR_FOREGROUND,
-                                            )
-                                            .clone(),
-                                    )
-                                    .max_width(
-                                        self.editor.size.borrow().width - 20.0,
-                                    )
-                                    .build()
-                                    .unwrap();
-                                text_height += 10.0 + text_layout.size().height;
-                                text_layout
-                            })
-                            .collect::<Vec<PietTextLayout>>()
-                    })
-                    .unwrap_or_else(Vec::new);
-
-                let start = diagnostic.diagnositc.range.start;
-                let rect = Rect::ZERO
-                    .with_origin(Point::new(
-                        0.0,
-                        (start.line + 1) as f64 * line_height,
-                    ))
-                    .with_size(Size::new(
-                        self.editor.size.borrow().width,
-                        text_height + 20.0,
-                    ));
-                ctx.fill(
-                    rect,
-                    self.config
-                        .get_color_unchecked(LapceTheme::EDITOR_SELECTION),
-                );
-
-                let severity = diagnostic
-                    .diagnositc
-                    .severity
-                    .as_ref()
-                    .unwrap_or(&DiagnosticSeverity::Information);
-                let color = match severity {
-                    DiagnosticSeverity::Error => {
-                        self.config.get_color_unchecked(LapceTheme::LAPCE_ERROR)
-                    }
-                    DiagnosticSeverity::Warning => {
-                        self.config.get_color_unchecked(LapceTheme::LAPCE_WARN)
-                    }
-                    _ => self.config.get_color_unchecked(LapceTheme::LAPCE_WARN),
-                };
-                ctx.stroke(rect, color, 1.0);
-                ctx.draw_text(
-                    &text_layout,
-                    Point::new(
-                        10.0 + self.editor.scroll_offset.x,
-                        (start.line + 1) as f64 * line_height + 10.0,
-                    ),
-                );
-                let mut text_height = text_size.height;
-
-                for text in related {
-                    text_height += 10.0;
-                    ctx.draw_text(
-                        &text,
-                        Point::new(
-                            10.0 + self.editor.scroll_offset.x,
-                            (start.line + 1) as f64 * line_height
-                                + 10.0
-                                + text_height,
-                        ),
-                    );
-                    text_height += text.size().height;
-                }
-            }
-        }
     }
 }
 
@@ -4762,93 +3459,6 @@ pub struct TabRect {
     pub text_layout: PietTextLayout,
 }
 
-impl TabRect {
-    pub fn paint(
-        &self,
-        ctx: &mut PaintCtx,
-        data: &LapceTabData,
-        widget_id: WidgetId,
-        i: usize,
-        size: Size,
-        mouse_pos: Point,
-    ) {
-        let width = 13.0;
-        let height = 13.0;
-        let editor_tab = data.main_split.editor_tabs.get(&widget_id).unwrap();
-
-        let rect = Size::new(width, height).to_rect().with_origin(Point::new(
-            self.rect.x0 + (size.height - width) / 2.0,
-            (size.height - height) / 2.0,
-        ));
-        if i == editor_tab.active {
-            ctx.fill(
-                self.rect,
-                data.config
-                    .get_color_unchecked(LapceTheme::EDITOR_BACKGROUND),
-            );
-        }
-        ctx.draw_svg(&self.svg, rect, None);
-        let text_size = self.text_layout.size();
-        ctx.draw_text(
-            &self.text_layout,
-            Point::new(
-                self.rect.x0 + size.height,
-                (size.height - text_size.height) / 2.0,
-            ),
-        );
-        let x = self.rect.x1;
-        ctx.stroke(
-            Line::new(Point::new(x - 0.5, 0.0), Point::new(x - 0.5, size.height)),
-            data.config.get_color_unchecked(LapceTheme::LAPCE_BORDER),
-            1.0,
-        );
-
-        if ctx.is_hot() {
-            if self.close_rect.contains(mouse_pos) {
-                ctx.fill(
-                    &self.close_rect,
-                    data.config
-                        .get_color_unchecked(LapceTheme::EDITOR_CURRENT_LINE),
-                );
-            }
-            if self.rect.contains(mouse_pos) {
-                let svg = get_svg("close.svg").unwrap();
-                ctx.draw_svg(
-                    &svg,
-                    self.close_rect.inflate(-4.0, -4.0),
-                    Some(
-                        data.config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND),
-                    ),
-                );
-            }
-        }
-
-        // Only display dirty icon if focus is not on tab bar, so that the close svg can be shown
-        if !(ctx.is_hot() && self.rect.contains(mouse_pos)) {
-            // See if any of the children are dirty
-            let is_dirty = match &editor_tab.children[i] {
-                EditorTabChild::Editor(editor_id, _) => {
-                    let buffer = data.main_split.editor_buffer(*editor_id);
-                    buffer.dirty
-                }
-            };
-
-            if is_dirty {
-                let svg = get_svg("unsaved.svg").unwrap();
-                ctx.draw_svg(
-                    &svg,
-                    self.close_rect.inflate(-4.0, -4.0),
-                    Some(
-                        data.config
-                            .get_color_unchecked(LapceTheme::EDITOR_FOREGROUND),
-                    ),
-                )
-            }
-        }
-    }
-}
-
 #[derive(Clone)]
 pub struct RegisterContent {
     #[allow(dead_code)]
@@ -4972,27 +3582,4 @@ fn process_get_references(
         Target::Auto,
     );
     Ok(())
-}
-
-fn paint_wave_line(
-    ctx: &mut PaintCtx,
-    origin: Point,
-    max_width: f64,
-    color: &Color,
-) {
-    let mut path = BezPath::new();
-    let mut x = 0.0;
-    let width = 3.5;
-    let height = 4.0;
-    path.move_to(origin + (0.0, height / 2.0));
-    let mut direction = 1.0;
-    while x < max_width {
-        let point = origin + (x, height / 2.0);
-        let p1 = point + (width / 2.0, -height / 2.0 * direction);
-        let p2 = point + (width, 0.0);
-        path.quad_to(p1, p2);
-        x += width;
-        direction *= -1.0;
-    }
-    ctx.stroke(path, color, 1.4);
 }
